+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Include mtime in job events
+- Job Events: Allow clients to filter events based on their action type.
 
 ### Changed